+++ tests/test_backend.py
@@ -9,23 +9,6 @@ import audbackend
 import audeer
 
 
-@pytest.fixture(scope='function')
-def backend(request):
-
-    name = request.param
-    host = pytest.HOSTS[name]
-    repository = f'unittest-{audeer.uid()[:8]}'
-    backend = audbackend.create(name, host, repository)
-
-    yield backend
-
-    # TODO: replace with audbackend.delete() when available
-    if name == 'artifactory':
-        backend._repo.delete()
-    elif name == 'file-system':
-        audeer.rmdir(backend._root)
-
-
 @pytest.mark.parametrize(
     'files, name, folder, version, tmp_root',
     [
@@ -137,57 +120,6 @@ def test_archive(tmpdir, files, name, folder, version, tmp_root, backend):
         version,
         tmp_root=tmp_root,
     ) == files_as_list
-
-
-@pytest.mark.parametrize(
-    'name, host, repository, cls',
-    [
-        (
-            'file-system',
-            pytest.HOSTS['file-system'],
-            f'unittest-{audeer.uid()[:8]}',
-            audbackend.FileSystem,
-        ),
-        (
-            'artifactory',
-            pytest.HOSTS['artifactory'],
-            f'unittest-{audeer.uid()[:8]}',
-            audbackend.Artifactory,
-        ),
-        pytest.param(  # backend does not exist
-            'bad-backend',
-            None,
-            None,
-            None,
-            marks=pytest.mark.xfail(raises=ValueError),
-        ),
-        pytest.param(  # host does not exist
-            'artifactory',
-            'bad-host',
-            'repo',
-            None,
-            marks=pytest.mark.xfail(raises=audbackend.BackendError),
-        ),
-        pytest.param(  # invalid repository name
-            'artifactory',
-            pytest.HOSTS['artifactory'],
-            'bad/repo',
-            None,
-            marks=pytest.mark.xfail(raises=audbackend.BackendError),
-        ),
-    ]
-)
-def test_create(name, host, repository, cls):
-
-    backend = audbackend.create(name, host, repository)
-
-    assert isinstance(backend, cls)
-
-    # TODO: replace with audbackend.delete() when available
-    if name == 'artifactory':
-        backend._repo.delete()
-    else:
-        audeer.rmdir(backend._root)
 
 
 @pytest.mark.parametrize(
